Revert "Add pdk group to Trusty dependencies"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7d9cf975-487c-44a4-8da2-1b49bd84d175). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377833261
Change-Id: Ie5d0b02c92a3991a246bc24c0b6e22a8d4c41da1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com