Revert "Remove 40 Rust crates."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c735b902-1385-4c6d-baba-09f8f3dc194f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d5d3dba1a329eadb8509a2d6ca25db2e1fce7ea
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com